+++ tests/arenagen_golden/editions2023.rp.hpp
@@ -19,6 +19,8 @@ class M {
  public:
   enum class Inner : std::int32_t {
     ZERO = 0,
+    rp_known_min = 0,
+    rp_known_max = 0,
     rp_non_exhaustive_min = INT32_MIN,
     rp_non_exhaustive_max = INT32_MAX,
   };