+++ tests/arenagen_golden/xref_prefixed/xref.rp.hpp
@@ -59,6 +59,8 @@ class Nested {
     enum class Kind : std::int32_t {
       UNKNOWN = 0,
       A = 1,
+      rp_known_min = 0,
+      rp_known_max = 1,
       rp_non_exhaustive_min = INT32_MIN,
       rp_non_exhaustive_max = INT32_MAX,
     };
@@ -268,6 +270,8 @@ class CousinB {
   enum class Tag : std::int32_t {
     NONE = 0,
     X = 1,
+    rp_known_min = 0,
+    rp_known_max = 1,
     rp_non_exhaustive_min = INT32_MIN,
     rp_non_exhaustive_max = INT32_MAX,
   };