+++ tests/common_golden/naming.rp.common.hpp
@@ -9,6 +9,8 @@ namespace nm {
 enum class E : std::int32_t {
   decode_ = 0,
   decode__ = 1,
+  rp_known_min = 0,
+  rp_known_max = 1,
   rp_non_exhaustive_min = INT32_MIN,
   rp_non_exhaustive_max = INT32_MAX,
 };