+++ tests/streamgen_golden/xref.rp.common.hpp
@@ -9,6 +9,8 @@ namespace xr {
 enum class Version : std::int32_t {
   VERSION_1 = 0,
   VERSION_2 = 1,
+  rp_known_min = 0,
+  rp_known_max = 1,
   rp_non_exhaustive_min = INT32_MIN,
   rp_non_exhaustive_max = INT32_MAX,
 };
@@ -16,6 +18,8 @@ enum class Version : std::int32_t {
 enum class Mixed : std::int32_t {
   MIXED_A = 0,
   OTHER = 1,
+  rp_known_min = 0,
+  rp_known_max = 1,
   rp_non_exhaustive_min = INT32_MIN,
   rp_non_exhaustive_max = INT32_MAX,
 };
@@ -23,6 +27,8 @@ enum class Mixed : std::int32_t {
 enum class Stream : std::int32_t {
   STREAM_OK = 0,
   STREAM_EOF = 1,
+  rp_known_min = 0,
+  rp_known_max = 1,
   rp_non_exhaustive_min = INT32_MIN,
   rp_non_exhaustive_max = INT32_MAX,
 };